+++ backend/osrs-dmg-sim/dpscalc/maxhit.go
@@ -242,7 +242,7 @@ func getMagicMaxHit(player *player) int { //TODO maybe look over again and have 
 			dpsDetailEntries.TrackValue(dpsdetail.DamageLevelPrayer, prayerStr)
 		}
 	}
-	magicDmgBonus := player.equipmentStats.damageStats.magicStrength + prayerStr
+	magicDmgBonus := player.equipmentStats.damageStats.magicStrength + float32(prayerStr)
 
 	gearMagicBonus := 0
 	if player.equippedGear.isAnyEquipped(smokeBattleStaves) && spell.spellbook == standardSpellBook {
@@ -265,9 +265,9 @@ func getMagicMaxHit(player *player) int { //TODO maybe look over again and have 
 			baseMaxhit = int(baseMaxhit * 23 / 20)
 		}
 		maxHit = int(baseMaxhit * (1000.0 + (gearMagicBonus)) / 1000.00)
-		maxHit = int(maxHit * (1000.0 + (magicDmgBonus)) / 1000.0)
+		maxHit = int(float32(maxHit) * (1000.0 + (magicDmgBonus)) / 1000.0)
 	} else {
-		maxHit = int(baseMaxhit * (1000.0 + (magicDmgBonus + gearMagicBonus)) / 1000.0)
+		maxHit = int(float32(baseMaxhit) * (1000.0 + (magicDmgBonus + float32(gearMagicBonus))) / 1000.0)
 		if blackMaskBonus {
 			maxHit = int(maxHit * 23 / 20)
 		}
